# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BanHang Backend API
  version: 2.0.0
  description: Versioned NestJS backend API for auth, CRM, content, catalog, cart, orders, payments, inventory, shipping, notifications, reporting, and store locator flows.
servers:
  - url: http://localhost:4000/api/v1
...
paths:
...
  /payments/saved-methods:
    get:
      tags: [Payments]
      summary: List saved payment methods for the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Saved payment methods
    post:
      tags: [Payments]
      summary: Save a reusable payment method token or card reference
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSavedPaymentMethodRequest'
      responses:
        '201':
          description: Saved payment method created
  /payments/saved-methods/{id}/default:
    post:
      tags: [Payments]
      summary: Mark one saved payment method as default
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Saved payment method updated
  /payments/saved-methods/{id}:
    delete:
      tags: [Payments]
      summary: Remove a saved payment method
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Saved payment method removed
  /payments:
    post:
      tags: [Payments]
      summary: Initiate payment for an order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiatePaymentRequest'
      responses:
        '200':
          description: Payment session created
  /payments/{id}:
    get:
      tags: [Payments]
      summary: Get payment status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment detail
  /payments/{id}/refund:
    post:
      tags: [Payments]
      summary: Refund payment fully or partially
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundPaymentRequest'
      responses:
        '200':
          description: Payment refunded
  /payments/webhook/{gateway}:
    post:
      tags: [Payments]
      summary: Gateway webhook with HMAC signature and idempotency
      parameters:
        - in: path
          name: gateway
          required: true
          schema:
            type: string
        - in: header
          name: x-webhook-signature
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentWebhookRequest'
      responses:
        '200':
          description: Webhook processed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    InitiatePaymentRequest:
      type: object
      required: [orderId]
      properties:
        orderId:
          type: string
        method:
          type: string
          enum: [cod, vnpay, momo, zalopay, stripe, paypal, bank_transfer]
        savedPaymentMethodId:
          type: string
        returnUrl:
          type: string
          format: uri
        ipAddress:
          type: string
        locale:
          type: string
    CreateSavedPaymentMethodRequest:
      type: object
      required: [gateway, method, label, tokenRef]
      properties:
        gateway:
          type: string
        method:
          type: string
          enum: [cod, vnpay, momo, zalopay, stripe, paypal, bank_transfer]
        label:
          type: string
        brand:
          type: string
        last4:
          type: string
        expiryMonth:
          type: integer
          minimum: 1
          maximum: 12
        expiryYear:
          type: integer
          minimum: 2000
        tokenRef:
          type: string
        providerCustomerRef:
          type: string
        isDefault:
          type: boolean
    RefundPaymentRequest:
      type: object
      properties:
        amount:
          type: integer
          minimum: 1
    PaymentWebhookRequest:
      type: object
      required: [eventId, type, payload]
      properties:
        eventId:
          type: string
        type:
          type: string
        orderId:
          type: string
        paymentId:
          type: string
        transactionId:
          type: string
        payload:
          type: object
          additionalProperties: true
...